requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
dateparser==1.2.0
Django==5.0.2
django-environ==0.11.2
environ==1.0
feedparser==6.0.11
idna==3.6
python-dateutil==2.8.2
pytz==2024.1
regex==2023.12.25
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
tzlocal==5.2
urllib3==2.2.0
